working resnet50 with hparams
