--- benchmarks/resnet50_i1k/__init__.py
@@ -1,12 +1,0 @@
-# Copyright 2021 MosaicML. All Rights Reserved.
-
-from composer.models.resnet50.model import ResNet50 as ResNet50
-from composer.models.resnet50.resnet50_hparams import ResNet50Hparams as ResNet50Hparams
-
-_task = 'Image Classification'
-_dataset = 'ImageNet'
-_name = 'ResNet50'
-_quality = '76.51'
-_metric = 'Top-1 Accuracy'
-_ttt = '3h 33m'
-_hparams = 'resnet50.yaml'
