+++ benchmarks/resnet50_i1k/resnet50.yaml
@@ -46,7 +46,7 @@ train_batch_size: 2048
 eval_batch_size: 2048
 seed: 17
 device:
-  gpu: {}
+  cpu: {}
 dataloader:
   pin_memory: true
   timeout: 0
